Change review_share type

+++ libraries/app/include/deip/app/deip_api_objects.hpp
@@ -518,7 +518,7 @@ struct research_api_obj
     std::string permlink;
     bool is_finished;
     share_type owned_tokens;
-    double review_share_in_percent;
+    uint16_t review_share_in_percent;
     time_point_sec created_at;
 };
